+++ samples/dotnet/kernel-syntax-examples/Example17_ChatGPT.cs
@@ -15,7 +15,12 @@ using RepoUtils;
 // ReSharper disable CommentTypo
 public static class Example17_ChatGPT
 {
-    public static async Task RunAsync()
+    public static async Task  RunAsync()
+    {
+        await OpenAIChatAsync();
+        await AzureChatAsync();
+    }
+    public static async Task OpenAIChatAsync()
     {
         Console.WriteLine("======== SK with ChatGPT ========");
 
@@ -74,6 +79,68 @@ public static class Example17_ChatGPT
         5. "The Histories" by Herodotus - This is a comprehensive account of the Persian Wars and provides a wealth of information about ancient Greek culture and society.
 
         I hope these suggestions are helpful!
+        ------------------------
+        */
+    }
+    public static async Task AzureChatAsync()
+    {
+        Console.WriteLine("======== SK with ChatGPT ========");
+
+        IKernel kernel = new KernelBuilder().WithLogger(ConsoleLogger.Log).Build();
+
+        // Add your chat completion service
+        kernel.Config.AddAzureChatCompletionService("chat", "gpt-35-turbo", Env.Var("AZURE_ENDPOINT"), Env.Var("AZURE_API_KEY"));
+
+        IChatCompletion chatGPT = kernel.GetService<IChatCompletion>();
+        var chat = (OpenAIChatHistory)chatGPT.CreateNewChat("You are a librarian, expert about books");
+
+        // First user message
+        chat.AddUserMessage("Hi, I'm looking for book suggestions");
+
+        // First bot message
+        string reply = await chatGPT.GenerateMessageAsync(chat, new ChatRequestSettings());
+        chat.AddAssistantMessage(reply);
+
+        // Second user message
+        chat.AddUserMessage("I love history and philosophy, I'd like to learn something new about Greece, any suggestion?");
+
+        // Second bot message
+        reply = await chatGPT.GenerateMessageAsync(chat, new ChatRequestSettings());
+        chat.AddAssistantMessage(reply);
+
+        Console.WriteLine("Chat content:");
+        Console.WriteLine("------------------------");
+        foreach (var message in chat.Messages)
+        {
+            Console.WriteLine($"{message.AuthorRole}: {message.Content}");
+            Console.WriteLine("------------------------");
+        }
+
+        /* Output:
+
+        Chat content:
+        ------------------------
+        system: You are a librarian, expert about books
+        ------------------------
+        user: Hi, I'm looking for book suggestions
+        ------------------------
+        assistant: Sure, I'd be happy to help! What kind of books are you interested in? Fiction or non-fiction? Any particular genre?
+        ------------------------
+        user: I love history and philosophy, I'd like to learn something new about Greece, any suggestion?
+        ------------------------
+        assistant: Great! For history and philosophy books about Greece, here are a few suggestions:
+
+        1. "The Greeks" by H.D.F. Kitto - This book provides an overview of ancient Greek civilization, including their philosophy, literature, and politics.
+
+        2. "The Peloponnesian War" by Thucydides - This is a classic work of history that chronicles the war between Athens and Sparta in the 5th century BC.
+
+        3. "The Republic" by Plato - This is one of the most famous works of philosophy in Western history, and it explores the nature of justice and the ideal society.
+
+        4. "The Iliad" by Homer - This epic poem tells the story of the Trojan War and is a great introduction to ancient Greek literature.
+
+        5. "The Histories" by Herodotus - This is another classic work of history that covers the Persian Wars and other events in ancient Greece.
+
+        I hope these suggestions help! Let me know if you have any other questions.
         ------------------------
         */
     }
